+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 ------------------
 
 - Updated Electron version from v2 to v8.
-- Default engine changed from `phantomjs` to `electron`.
-- `SlimerJS` is deprecated.
+- Removed support for `PhantomJS` and `SlimerJS` (engines are outdated).
 
 
 3.1.0 / 2017-09-22